Silence 4.6.2 compile (no whatsnew)
git-svn-id: svn://messdev.no-ip.org/mess@13431 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/tools/imgtool/windows/opcntrl.c
@@ -190,7 +190,7 @@ static BOOL check_editbox(HWND control)
 	optreserr_t err;
 	HANDLE h;
 	void *val;
-	int i_val;
+	UINT64 i_val;
 
 
 	guide = (const option_guide *) GetProp(control, guide_prop);
